+++ src/Http/Response.php
@@ -1,7 +1,7 @@
 <?php
 namespace Phly\Conduit\Http;
 
-use Psr\Http\Message\OutgoingResponseInterface;
+use Psr\Http\Message\ResponseInterface as PsrResponseInterface;
 use Psr\Http\Message\StreamableInterface;
 
 /**
@@ -11,7 +11,7 @@ use Psr\Http\Message\StreamableInterface;
  * to provide a common interface for all PSR HTTP implementations.
  */
 class Response implements
-    OutgoingResponseInterface,
+    PsrResponseInterface,
     ResponseInterface
 {
     /**
@@ -20,14 +20,14 @@ class Response implements
     private $complete = false;
 
     /**
-     * @var BaseResponseInterface
+     * @var PsrResponseInterface
      */
     private $psrResponse;
 
     /**
-     * @param OutgoingResponseInterface $response
+     * @param PsrResponseInterface $response
      */
-    public function __construct(OutgoingResponseInterface $response)
+    public function __construct(PsrResponseInterface $response)
     {
         $this->psrResponse = $response;
     }
@@ -35,7 +35,7 @@ class Response implements
     /**
      * Return the original PSR response object
      *
-     * @return BaseResponseInterface
+     * @return PsrResponseInterface
      */
     public function getOriginalResponse()
     {
@@ -52,10 +52,11 @@ class Response implements
     public function write($data)
     {
         if ($this->complete) {
-            return;
+            return $this;
         }
 
         $this->getBody()->write($data);
+        return $this;
     }
 
     /**
@@ -72,14 +73,16 @@ class Response implements
     public function end($data = null)
     {
         if ($this->complete) {
-            return;
+            return $this;
         }
 
         if ($data) {
             $this->write($data);
         }
 
-        $this->complete = true;
+        $new = clone $this;
+        $new->complete = true;
+        return $new;
     }
 
     /**
@@ -95,7 +98,7 @@ class Response implements
     }
 
     /**
-     * Proxy to BaseResponseInterface::getProtocolVersion()
+     * Proxy to PsrResponseInterface::getProtocolVersion()
      *
      * @return string HTTP protocol version.
      */
@@ -105,18 +108,19 @@ class Response implements
     }
 
     /**
-     * Proxy to BaseResponseInterface::setProtocolVersion()
-     * 
-     * @param string $version 
-     * @return void
+     * Proxy to PsrResponseInterface::withProtocolVersion()
+     *
+     * @param string $version
+     * @return Response
      */
-    public function setProtocolVersion($version)
+    public function withProtocolVersion($version)
     {
-        return $this->psrResponse->setProtocolVersion($version);
+        $new = $this->psrResponse->withProtocolVersion($version);
+        return new self($new);
     }
 
     /**
-     * Proxy to BaseResponseInterface::getBody()
+     * Proxy to PsrResponseInterface::getBody()
      *
      * @return StreamableInterface|null Returns the body, or null if not set.
      */
@@ -126,22 +130,24 @@ class Response implements
     }
 
     /**
-     * Proxy to BaseResponseInterface::setBody()
+     * Proxy to PsrResponseInterface::withBody()
      *
      * @param StreamableInterface $body Body.
+     * @return Response
      * @throws \InvalidArgumentException When the body is not valid.
      */
-    public function setBody(StreamableInterface $body)
+    public function withBody(StreamableInterface $body)
     {
         if ($this->complete) {
-            return;
+            return $this;
         }
 
-        return $this->psrResponse->setBody($body);
+        $new = $this->psrResponse->withBody($body);
+        return new self($new);
     }
 
     /**
-     * Proxy to BaseResponseInterface::getHeaders()
+     * Proxy to PsrResponseInterface::getHeaders()
      *
      * @return array Returns an associative array of the message's headers.
      */
@@ -151,7 +157,7 @@ class Response implements
     }
 
     /**
-     * Proxy to BaseResponseInterface::hasHeader()
+     * Proxy to PsrResponseInterface::hasHeader()
      *
      * @param string $header Case-insensitive header name.
      * @return bool Returns true if any header names match the given header
@@ -164,7 +170,7 @@ class Response implements
     }
 
     /**
-     * Proxy to BaseResponseInterface::getHeader()
+     * Proxy to PsrResponseInterface::getHeader()
      *
      * @param string $header Case-insensitive header name.
      * @return string
@@ -175,62 +181,68 @@ class Response implements
     }
 
     /**
-     * Proxy to BaseResponseInterface::getHeaderAsArray()
+     * Proxy to PsrResponseInterface::getHeaderAsArray()
      *
      * @param string $header Case-insensitive header name.
      * @return string[]
      */
-    public function getHeaderAsArray($header)
+    public function getHeaderLines($header)
     {
-        return $this->psrResponse->getHeaderAsArray($header);
+        return $this->psrResponse->getHeaderLines($header);
     }
 
     /**
-     * Proxy to BaseResponseInterface::setHeader()
+     * Proxy to PsrResponseInterface::withHeader()
      *
      * @param string $header Header name
      * @param string|string[] $value  Header value(s)
+     * @return Response
      */
-    public function setHeader($header, $value)
+    public function withHeader($header, $value)
     {
         if ($this->complete) {
-            return;
+            return $this;
         }
 
-        return $this->psrResponse->setHeader($header, $value);
+        $new = $this->psrResponse->withHeader($header, $value);
+        return new self($new);
     }
 
     /**
-     * Proxy to BaseResponseInterface::addHeader()
+     * Proxy to PsrResponseInterface::withAddedHeader()
      *
      * @param string $header Header name to add or append
      * @param string|string[] $value Value(s) to add or merge into the header
+     * @return Response
      */
-    public function addHeader($header, $value)
+    public function withAddedHeader($header, $value)
     {
         if ($this->complete) {
-            return;
+            return $this;
         }
 
-        return $this->psrResponse->addHeader($header, $value);
+        $new = $this->psrResponse->withAddedHeader($header, $value);
+        return new self($new);
     }
 
     /**
-     * Proxy to BaseResponseInterface::removeHeader()
+     * Proxy to PsrResponseInterface::withoutHeader()
      *
      * @param string $header HTTP header to remove
+     * @return Response
      */
-    public function removeHeader($header)
+    public function withoutHeader($header)
     {
         if ($this->complete) {
-            return;
+            return $this;
         }
 
-        return $this->psrResponse->removeHeader($header);
+        $new = $this->psrResponse->withoutHeader($header);
+        return new self($new);
     }
 
     /**
-     * Proxy to BaseResponseInterface::getStatusCode()
+     * Proxy to PsrResponseInterface::getStatusCode()
      *
      * @return integer Status code.
      */
@@ -240,22 +252,24 @@ class Response implements
     }
 
     /**
-     * Proxy to BaseResponseInterface::setStatus()
+     * Proxy to PsrResponseInterface::withStatus()
      *
      * @param integer $code The 3-digit integer result code to set.
      * @param null|string $reasonPhrase The reason phrase to use with the status, if any.
+     * @return Response
      */
-    public function setStatus($code, $reasonPhrase = null)
+    public function withStatus($code, $reasonPhrase = null)
     {
         if ($this->complete) {
-            return;
+            return $this;
         }
 
-        return $this->psrResponse->setStatus($code, $reasonPhrase);
+        $new = $this->psrResponse->withStatus($code, $reasonPhrase);
+        return new self($new);
     }
 
     /**
-     * Proxy to BaseResponseInterface::getReasonPhrase()
+     * Proxy to PsrResponseInterface::getReasonPhrase()
      *
      * @return string|null Reason phrase, or null if unknown.
      */