+++ test/Http/RequestTest.php
@@ -2,7 +2,7 @@
 namespace PhlyTest\Conduit\Http;
 
 use Phly\Conduit\Http\Request;
-use Phly\Http\IncomingRequest as PsrRequest;
+use Phly\Http\ServerRequest as PsrRequest;
 use Phly\Http\Uri;
 use PHPUnit_Framework_TestCase as TestCase;
 
@@ -10,56 +10,42 @@ class RequestTest extends TestCase
 {
     public function setUp()
     {
-        $this->original = new PsrRequest(
-            'http://example.com/',
-            'GET',
-            [],
-            'php://memory'
-        );
+        $psrRequest     = new PsrRequest([], [], 'http://example.com/', 'GET', 'php://memory');
+        $this->original = $psrRequest;
         $this->request  = new Request($this->original);
     }
 
-    public function testAllowsManipulatingArbitraryNonPrivateProperties()
-    {
-        $this->request->originalUrl = 'http://foo.example.com/foo';
-        $this->assertTrue(isset($this->request->originalUrl));
-        $this->assertEquals('http://foo.example.com/foo', $this->request->originalUrl);
-        unset($this->request->originalUrl);
-        $this->assertNull($this->request->originalUrl);
-    }
-
-    public function testFetchingUnknownPropertyYieldsNull()
-    {
-        $this->assertNull($this->request->somePropertyWeMadeUp);
-    }
-
-    public function testArrayPropertyValueIsCastToArrayObject()
-    {
-        $original = ['test' => 'value'];
-        $this->request->anArray = $original;
-        $this->assertInstanceOf('ArrayObject', $this->request->anArray);
-        $this->assertEquals($original, $this->request->anArray->getArrayCopy());
-    }
-
-    public function testCallingSetUrlSetsOriginalUrlProperty()
+    public function testCallingSetUriSetsUriInRequestAndOriginalRequestInClone()
     {
         $url = 'http://example.com/foo';
-        $this->request->setUrl($url);
-        $this->assertSame('http://example.com/', $this->request->originalUrl);
-        $this->assertSame($url, $this->request->getUrl());
+        $request = $this->request->withUri(new Uri($url));
+        $this->assertNotSame($this->request, $request);
+        $this->assertSame($this->original, $request->getOriginalRequest());
+        $this->assertSame($url, (string) $request->getUri());
     }
 
-    public function testConstructorSetsOriginalUrlIfDecoratedRequestHasUrl()
+    public function testConstructorSetsOriginalRequestIfNoneProvided()
     {
         $url = 'http://example.com/foo';
-        $baseRequest = new PsrRequest(
-            $url,
-            'GET',
-            [],
-            'php://memory'
-        );
+        $baseRequest = new PsrRequest([], [], $url, 'GET', 'php://memory');
+
         $request = new Request($baseRequest);
-        $this->assertSame($baseRequest->getUrl(), $request->originalUrl);
+        $this->assertSame($baseRequest, $request->getOriginalRequest());
+    }
+
+    public function testCallingSettersRetainsOriginalRequest()
+    {
+        $url = 'http://example.com/foo';
+        $baseRequest = new PsrRequest([], [], $url, 'GET', 'php://memory');
+
+        $request = new Request($baseRequest);
+        $request = $request->withMethod('POST');
+        $new     = $request->withAddedHeader('X-Foo', 'Bar');
+
+        $this->assertNotSame($request, $new);
+        $this->assertNotSame($baseRequest, $new);
+        $this->assertNotSame($baseRequest, $new->getCurrentRequest());
+        $this->assertSame($baseRequest, $new->getOriginalRequest());
     }
 
     public function testCanAccessOriginalRequest()
@@ -70,35 +56,14 @@ class RequestTest extends TestCase
     public function testDecoratorProxiesToAllMethods()
     {
         $stream = $this->getMock('Psr\Http\Message\StreamableInterface');
-        $psrRequest = new PsrRequest(
-            'http://example.com/',
-            'POST',
-            [
-                'Accept' => 'application/xml',
-                'X-URL'  => 'http://example.com/foo',
-            ],
-            $stream
-        );
+        $psrRequest = new PsrRequest([], [], 'http://example.com', 'POST', $stream, [
+            'Accept' => 'application/xml',
+            'X-URL' => 'http://example.com/foo',
+        ]);
         $request = new Request($psrRequest);
 
         $this->assertEquals('1.1', $request->getProtocolVersion());
         $this->assertSame($stream, $request->getBody());
         $this->assertSame($psrRequest->getHeaders(), $request->getHeaders());
-    }
-
-    public function testPropertyAccessProxiesToRequestAttributes()
-    {
-        $this->original->setAttributes([
-            'foo' => 'bar',
-            'bar' => 'baz',
-        ]);
-
-        $this->assertTrue(isset($this->request->foo));
-        $this->assertTrue(isset($this->request->bar));
-        $this->assertFalse(isset($this->request->baz));
-
-        $this->request->baz = 'quz';
-        $this->assertTrue(isset($this->request->baz));
-        $this->assertEquals('quz', $this->original->getAttribute('baz', false));
     }
 }