+++ test/MiddlewareTest.php
@@ -5,8 +5,9 @@ use Phly\Conduit\Http\Request as RequestDecorator;
 use Phly\Conduit\Http\Response as ResponseDecorator;
 use Phly\Conduit\Middleware;
 use Phly\Conduit\Utils;
-use Phly\Http\IncomingRequest as Request;
-use Phly\Http\OutgoingResponse as Response;
+use Phly\Http\ServerRequest as Request;
+use Phly\Http\Response;
+use Phly\Http\Uri;
 use PHPUnit_Framework_TestCase as TestCase;
 use ReflectionProperty;
 
@@ -14,12 +15,7 @@ class MiddlewareTest extends TestCase
 {
     public function setUp()
     {
-        $this->request    = new Request(
-            'http://example.com/',
-            'GET',
-            [],
-            'php://memory'
-        );
+        $this->request    = new Request([], [], 'http://example.com/', 'GET', 'php://memory');
         $this->response   = new Response();
         $this->middleware = new Middleware();
     }
@@ -64,7 +60,7 @@ class MiddlewareTest extends TestCase
             $phpunit->fail('Should not hit fourth handler!');
         });
 
-        $request = new Request('http://local.example.com/foo', 'GET', [], 'php://memory');
+        $request = new Request([], [], 'http://local.example.com/foo', 'GET', 'php://memory');
         $this->middleware->__invoke($request, $this->response);
         $body = (string) $this->response->getBody();
         $this->assertContains('First', $body);
@@ -92,7 +88,7 @@ class MiddlewareTest extends TestCase
             $phpunit->fail('Should not hit fourth handler!');
         });
 
-        $request = new Request('http://local.example.com/foo', 'GET', [], 'php://memory');
+        $request = new Request([], [], 'http://local.example.com/foo', 'GET', 'php://memory');
         $this->middleware->__invoke($request, $this->response);
         $body = (string) $this->response->getBody();
         $this->assertContains('First', $body);
@@ -117,7 +113,7 @@ class MiddlewareTest extends TestCase
             $next();
         });
 
-        $request = new Request('http://local.example.com/foo', 'GET', [], 'php://memory');
+        $request = new Request([], [], 'http://local.example.com/foo', 'GET', 'php://memory');
         $this->middleware->__invoke($request, $this->response, $out);
         $this->assertTrue($triggered);
     }
@@ -153,7 +149,7 @@ class MiddlewareTest extends TestCase
 
     public function testCanUseDecoratedRequestAndResponseDirectly()
     {
-        $baseRequest = new Request('http://local.example.com/foo', 'GET', [], 'php://memory');
+        $baseRequest = new Request([], [], 'http://local.example.com/foo', 'GET', 'php://memory');
 
         $request  = new RequestDecorator($baseRequest);
         $response = new ResponseDecorator($this->response);
@@ -172,5 +168,52 @@ class MiddlewareTest extends TestCase
         });
 
         $this->assertTrue($executed);
+    }
+
+    public function testReturnsOrigionalResponseIfStackDoesNotReturnAResponse()
+    {
+        $this->middleware->pipe(function ($req, $res, $next) {
+            $next();
+        });
+        $this->middleware->pipe(function ($req, $res, $next) {
+            $next();
+        });
+        $this->middleware->pipe(function ($req, $res, $next) {
+            return;
+        });
+        $phpunit = $this;
+        $this->middleware->pipe(function ($req, $res, $next) use ($phpunit) {
+            $phpunit->fail('Should not hit fourth handler!');
+        });
+
+        $request = new Request([], [], 'http://local.example.com/foo', 'GET', 'php://memory');
+        $result  = $this->middleware->__invoke($request, $this->response);
+        $this->assertSame($this->response, $result->getOriginalResponse());
+    }
+
+    public function testReturnsResponseReturnedByStack()
+    {
+        $return = new Response();
+
+        $this->middleware->pipe(function ($req, $res, $next) {
+            $next();
+        });
+        $this->middleware->pipe(function ($req, $res, $next) {
+            return $next();
+        });
+        $this->middleware->pipe(function ($req, $res, $next) use ($return) {
+            return $return;
+        });
+        $phpunit = $this;
+        $this->middleware->pipe(function ($req, $res, $next) use ($phpunit) {
+            $phpunit->fail('Should not hit fourth handler!');
+        });
+
+        $request = new Request([], [], 'http://local.example.com/foo', 'GET', 'php://memory');
+        $result  = $this->middleware->__invoke($request, $this->response);
+        $this->assertSame($return, $result, var_export([
+            spl_object_hash($return) => get_class($return),
+            spl_object_hash($result) => get_class($result),
+        ], 1));
     }
 }