+++ test/NextTest.php
@@ -6,21 +6,18 @@ use Phly\Conduit\Http\Request;
 use Phly\Conduit\Http\Response;
 use Phly\Conduit\Next;
 use Phly\Conduit\Route;
-use Phly\Http\IncomingRequest as PsrRequest;
-use Phly\Http\OutgoingResponse as PsrResponse;
+use Phly\Http\ServerRequest as PsrRequest;
+use Phly\Http\Response as PsrResponse;
+use Phly\Http\Uri;
 use PHPUnit_Framework_TestCase as TestCase;
 
 class NextTest extends TestCase
 {
     public function setUp()
     {
+        $psrRequest     = new PsrRequest([], [], 'http://example.com/', 'GET', 'php://memory');
         $this->stack    = new ArrayObject();
-        $this->request  = new Request(new PsrRequest(
-            'http://example.com/',
-            'GET',
-            [],
-            'php://memory'
-        ));
+        $this->request  = new Request($psrRequest);
         $this->response = new Response(new PsrResponse());
     }
 
@@ -51,7 +48,7 @@ class NextTest extends TestCase
             $triggered = true;
         };
 
-        $this->request->setUrl('http://local.example.com/bar');
+        $this->request->withUri(new Uri('http://local.example.com/bar'));
 
         $next = new Next($this->stack, $this->request, $this->response, $done);
         $next();
@@ -72,7 +69,7 @@ class NextTest extends TestCase
             $triggered = true;
         };
 
-        $this->request->setUrl('http://local.example.com/foobar');
+        $this->request->withUri(new Uri('http://local.example.com/foobar'));
 
         $next = new Next($this->stack, $this->request, $this->response, $done);
         $next();
@@ -93,9 +90,9 @@ class NextTest extends TestCase
             $phpunit->fail('Should not hit done handler');
         };
 
-        $this->request->setUrl('http://local.example.com/foo');
+        $request = $this->request->withUri(new Uri('http://local.example.com/foo'));
 
-        $next = new Next($this->stack, $this->request, $this->response, $done);
+        $next = new Next($this->stack, $request, $this->response, $done);
         $next();
         $this->assertTrue($triggered);
     }
@@ -106,7 +103,7 @@ class NextTest extends TestCase
         // then the URI path in the handler is "/bar"
         $triggered = null;
         $route = new Route('/foo', function ($req, $res, $next) use (&$triggered) {
-            $triggered = parse_url($req->getUrl(), PHP_URL_PATH);
+            $triggered = $req->getUri()->getPath();
         });
         $this->stack[] = $route;
 
@@ -115,9 +112,9 @@ class NextTest extends TestCase
             $phpunit->fail('Should not hit done handler');
         };
 
-        $this->request->setUrl('http://local.example.com/foo/bar');
+        $request = $this->request->withUri(new Uri('http://local.example.com/foo/bar'));
 
-        $next = new Next($this->stack, $this->request, $this->response, $done);
+        $next = new Next($this->stack, $request, $this->response, $done);
         $next();
         $this->assertEquals('/bar', $triggered);
     }
@@ -132,6 +129,7 @@ class NextTest extends TestCase
         });
         $route3 = new Route('/foo/baz', function ($req, $res, $next) {
             $res->end('done');
+            return $res;
         });
 
         $this->stack->append($route1);
@@ -143,9 +141,143 @@ class NextTest extends TestCase
             $phpunit->fail('Should not hit final handler');
         };
 
-        $this->request->setUrl('http://example.com/foo/baz/bat');
-        $next = new Next($this->stack, $this->request, $this->response, $done);
+        $request = $this->request->withUri(new Uri('http://example.com/foo/baz/bat'));
+        $next = new Next($this->stack, $request, $this->response, $done);
         $next();
         $this->assertEquals('done', (string) $this->response->getBody());
+    }
+
+    public function testMiddlewareReturningResponseShortcircuits()
+    {
+        $phpunit = $this;
+        $route1 = new Route('/foo', function ($req, $res, $next) {
+            return $res;
+        });
+        $route2 = new Route('/foo/bar', function ($req, $res, $next) use ($phpunit) {
+            $next();
+            $phpunit->fail('Should not hit route2 handler');
+        });
+        $route3 = new Route('/foo/baz', function ($req, $res, $next) use ($phpunit) {
+            $next();
+            $phpunit->fail('Should not hit route3 handler');
+        });
+
+        $this->stack->append($route1);
+        $this->stack->append($route2);
+        $this->stack->append($route3);
+
+        $done = function ($err) use ($phpunit) {
+            $phpunit->fail('Should not hit final handler');
+        };
+
+        $request = $this->request->withUri(new Uri('http://example.com/foo/bar/baz'));
+        $next = new Next($this->stack, $request, $this->response, $done);
+        $result = $next();
+        $this->assertSame($this->response, $result);
+    }
+
+    public function testMiddlewareCallingNextWithResponseAsFirstArgumentResetsResponse()
+    {
+        $phpunit = $this;
+        $cannedResponse = new Response(new PsrResponse());
+        $triggered = false;
+
+        $route1 = new Route('/foo', function ($req, $res, $next) use ($cannedResponse) {
+            return $next($cannedResponse);
+        });
+        $route2 = new Route('/foo/bar', function ($req, $res, $next) use ($phpunit, $cannedResponse, &$triggered) {
+            $phpunit->assertSame($cannedResponse, $res);
+            $triggered = true;
+        });
+
+        $this->stack->append($route1);
+        $this->stack->append($route2);
+
+        $done = function ($err) use ($phpunit) {
+            $phpunit->fail('Should not hit final handler');
+        };
+
+        $request = $this->request->withUri(new Uri('http://example.com/foo/bar/baz'));
+        $next = new Next($this->stack, $request, $this->response, $done);
+        $result = $next();
+        $this->assertTrue($triggered);
+        $this->assertSame($cannedResponse, $result);
+    }
+
+    public function testMiddlewareCallingNextWithRequestResetsRequest()
+    {
+        $phpunit       = $this;
+        $request       = $this->request->withUri(new Uri('http://example.com/foo/bar/baz'));
+        $cannedRequest = clone $request;
+        $cannedRequest = $cannedRequest->withMethod('POST');
+
+        $route1 = new Route('/foo/bar', function ($req, $res, $next) use ($cannedRequest) {
+            return $next($cannedRequest);
+        });
+        $route2 = new Route('/foo/bar/baz', function ($req, $res, $next) use ($phpunit, $cannedRequest) {
+            $phpunit->assertEquals($cannedRequest->getMethod(), $req->getMethod());
+            return $res;
+        });
+
+        $this->stack->append($route1);
+        $this->stack->append($route2);
+
+        $done = function ($err) use ($phpunit) {
+            $phpunit->fail('Should not hit final handler');
+        };
+
+        $next = new Next($this->stack, $request, $this->response, $done);
+        $next();
+    }
+
+    public function testMiddlewareCallingNextWithResponseResetsResponse()
+    {
+        $phpunit        = $this;
+        $cannedResponse = new Response(new PsrResponse());
+
+        $route1 = new Route('/foo', function ($req, $res, $next) use ($cannedResponse) {
+            return $next(null, $cannedResponse);
+        });
+        $route2 = new Route('/foo/bar', function ($req, $res, $next) use ($phpunit, $cannedResponse) {
+            $phpunit->assertSame($cannedResponse, $res);
+            return $res;
+        });
+
+        $this->stack->append($route1);
+        $this->stack->append($route2);
+
+        $done = function ($err) use ($phpunit) {
+            $phpunit->fail('Should not hit final handler');
+        };
+
+        $request = $this->request->withUri(new Uri('http://example.com/foo/bar/baz'));
+        $next = new Next($this->stack, $request, $this->response, $done);
+        $next();
+    }
+
+    public function testNextShouldReturnCurrentResponseAlways()
+    {
+        $phpunit        = $this;
+        $cannedResponse = new Response(new PsrResponse());
+
+        $route1 = new Route('/foo', function ($req, $res, $next) use ($cannedResponse) {
+            $next(null, $cannedResponse);
+        });
+        $route2 = new Route('/foo/bar', function ($req, $res, $next) use ($phpunit, $cannedResponse) {
+            $phpunit->assertSame($cannedResponse, $res);
+            return $res;
+        });
+
+        $this->stack->append($route1);
+        $this->stack->append($route2);
+
+        $done = function ($err) use ($phpunit) {
+            $phpunit->fail('Should not hit final handler');
+        };
+
+        $request = $this->request->withUri(new Uri('http://example.com/foo/bar/baz'));
+        $next    = new Next($this->stack, $request, $this->response, $done);
+        $result  = $next();
+        $this->assertSame($cannedResponse, $result);
     }
 }